Add Card and Deck Scripts v1.0
The Card class, for now, just has the public variables for name, type, attack, defense, effect1, and effect2.

The Deck class will generate a new Deck on start and draw through the deck. This needs the CardList text file to work. When adding a deck to an object, drag and drop the CardList text file into the cardStuff (to be renamed later) global variable.